moved messages that caused error
