+++ app/src/main/java/com/group7/dominion/Card/ImageButtonHandler.java
@@ -12,7 +12,7 @@ import com.floriankleewein.commonclasses.Cards.EstateType;
 import com.floriankleewein.commonclasses.Cards.MoneyCard;
 import com.floriankleewein.commonclasses.Cards.MoneyType;
 import com.floriankleewein.commonclasses.Network.ClientConnector;
-import com.floriankleewein.commonclasses.Network.Messages.GameLogicMsg.BuyCardMsg;
+import com.floriankleewein.commonclasses.Network.BuyCardMsg;
 import com.group7.dominion.R;
 
 import androidx.fragment.app.FragmentManager;
